GUI/requirements.txt
nbformat == 5.10.4
streamlit == 1.32.0 
scikit-learn == 1.4.2 
Pillow == 10.4.0
numpy == 1.26.0
pandas == 2.2.2
matplotlib == 3.8.4
seaborn == 0.13.2
openpyxl == 3.1.2
plotly == 5.23.0
nbconvert == 7.10.0